feat: added dynamic metadata
